+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/missingRio/missing-rio.routing.js
@@ -1,0 +1,15 @@
+import controller from './missing-rio.controller';
+import template from './missing-rio.html';
+
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state('telecom.packs.pack.xdsl.line.missing-rio', {
+    url: '/missingRio',
+    views: {
+      'accessView@telecom.packs.pack.xdsl.line': {
+        template,
+        controller,
+        controllerAs: 'PackXdslMissingRio',
+      },
+    },
+  });
+};
